Rename ManagePayees component and file
