+++ packages/desktop-client/src/components/payees/ManagePayeesWithData.js
@@ -7,7 +7,7 @@ import { applyChanges } from 'loot-core/src/shared/util';
 import { useActions } from '../../hooks/useActions';
 import useCategories from '../../hooks/useCategories';
 
-import { ManagePayees } from '.';
+import { ManagePayees } from './ManagePayees';
 
 export default function ManagePayeesWithData({ initialSelectedIds }) {
   let initialPayees = useSelector(state => state.queries.payees);
